# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ellowring API
  description: |
    Ellowring Software Solutions — Enterprise SaaS REST API (Phase 5).
    Learn. Prepare. Build. Get Hired.
  version: 1.0.0
  contact:
    name: Ellowring Engineering
    email: engineering@ellowring.com
servers:
  - url: http://localhost:4000/api
    description: Local
  - url: https://api-staging.ellowring.com/api
    description: Staging
  - url: https://api.ellowring.com/api
    description: Production
...
paths:
  /v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user and role profile; returns JWT access + refresh pair.
      operationId: post__auth_register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Email/password login; returns rotated token pair.
      operationId: post__auth_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/auth/otp/request:
    post:
      tags:
        - Authentication
      summary: Request OTP for passwordless login or verification.
      operationId: post__auth_otp_request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/otp/verify:
    post:
      tags:
        - Authentication
      summary: Verify OTP and issue token pair.
      operationId: post__auth_otp_verify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Start password reset; emails/SMS reset link or OTP.
      operationId: post__auth_forgot-password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Complete password reset with token.
      operationId: post__auth_reset-password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Rotate refresh token; issue new access + refresh.
      operationId: post__auth_refresh
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout; revoke refresh token (and optional all sessions).
      operationId: post__auth_logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get authenticated user profile with role extensions.
      operationId: get__auth_me
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password for authenticated user.
      operationId: post__auth_change-password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/auth/profile:
    patch:
      tags:
        - Authentication
      summary: "Update basic profile fields (name, phone, avatar)."
      operationId: patch__auth_profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data: {}
        meta:
          type: object
          properties:
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            pagination:
              $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        hasPrev: { type: boolean }